Create a_jump, a_mark

+++ script/sh/general_command/mv.sh
@@ -1,12 +1,12 @@
 if [ $# -ne 2 ]; then
-    echo 'usage : mv <source> <target>'
-    echo '    source and target are relative path from $shellscript'
+    echo 'usage : mv <source_file> <target_file>'
+    echo '  source and target are relative path from $shellscript'
     return 1
 fi
 
 
 
-source_absolute=$shellscript/$1
+source_absolute="$shellscript/$1"
 
 if [ ! -e "$source_absolute" ]; then
     echo 'a_command::mv : no file exist ['"$source_absolute"']'
@@ -25,30 +25,27 @@ fi
 
 
 
-소스와 타겟에 따라 다양한 경우가 존재할 수 있다
-1. 소스가 디렉토리이고 타겟이 디렉토리인 경우
-2. 소스가 디렉토리이고 타겟이 레귤러 파일인 경우
-3. 소스가 레귤러 파일이고 타겟이 디렉토리인 경우
-4. 소스가 레귤러 파일이고 타겟이 레귤러 파일인 경우 
+target_directory="$(dirname "$shellscript/$2")"
+if [ ! -e "$target_directory" ]; then
+    echo '[$shellscript/'"$(dirname $2)"'] is not exist'
+    unset target_directory
+    return 1
+fi
+unset target_directory
 
 
-mv "$shellscript/$1" "$shellscript/$2"
 
 source_basename=$(basename $1)
 source_filename=${source_basename%.*}
 
+rm -f "$bin/$source_filename"
+
+mv "$shellscript/$1" "$shellscript/$2"
+
 target_basename=$(basename $2)
 target_filename=${target_basename%.*}
 
-rm "$bin/$source_filename"
-
-if [ "$source_filename" = "$target_filename" ]; then
-    option='-s'
-else
-    option='-sf'
-fi
-
-ln "$option" "../script/sh/$2" "$bin/$target_filename"
+ln -s "../script/sh/$2" "$bin/$target_filename"
 
 unset source_basename
 unset source_filename
